+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fix `SessionRequestDetails.generate_file_request` handling of file paths with folders
 
+### Added
+- Add ability to delete/add cells with NoteableClient
+
+
 ## [0.0.2] - 2022-08-02
 ### Changed
 - Change `name` in pyproject.toml from `origami` to `noteable-origami`